feat: added average price endpoint

+++ BOB.Power/Models/ProxyModels/Proxy_PriceInterval.cs
@@ -1,9 +1,9 @@
 namespace BOB.Power.Models.ProxyModels;
 public class Proxy_PriceInterval
 {
-    public double NOK_per_kWh { get; set; }
-    public double EUR_per_kWh { get; set; }
-    public double EXR { get; set; }
+    public decimal NOK_per_kWh { get; set; }
+    public decimal EUR_per_kWh { get; set; }
+    public decimal EXR { get; set; }
     public DateTime time_start { get; set; }
     public DateTime time_end { get; set; }
 }
